mcp_server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.3",
    "asyncpg>=0.30",
    "playwright>=1.50",
    "httpx>=0.28",
]

[tool.hatch.build.targets.wheel]
packages = ["mcp_server"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
